SWARM-843: Wrong default port number for the legacy remoting connector
Motivation
----------
The remoting conector used wrong port for legacy systems

Modifications
-------------
Update the port number and java docs

Result
------
The legacy connector uses 4447 as the port number

+++ fractions/javaee/remoting/src/main/java/org/wildfly/swarm/remoting/RemotingFraction.java
@@ -63,5 +63,5 @@ public class RemotingFraction extends Remoting<RemotingFraction> implements Frac
     public int port() {
         return this.port.get();
     }
-    private Defaultable<Integer> port = integer(4777);
+    private Defaultable<Integer> port = integer(4447);
 }
